requirements.txt
streamlit
google-generativeai
gTTS
Pillow
pypdf